requirements.txt
langchain==0.3.21
langchain-community==0.3.20
langchain-core==0.3.48
langchain-text-splitters==0.3.7
langchainhub==0.1.21
langsmith==0.3.18
langchain-google-genai==2.0.6
langchain-openai==0.3.9
python-dotenv==1.0.1
urllib3==2.3.0
langchain-pinecone==0.2.3
streamlit==1.44.0